HBASE-30175 Upgrade hbase-server to use junit5 Part19 (#8278)
(cherry picked from commit 8d39f9f372d14c6c781789b6780cfcc88be6c7ec)
