+++ hbase-server/src/test/java/org/apache/hadoop/hbase/replication/regionserver/TestRegionReplicaReplicationEndpoint.java
@@ -17,11 +17,11 @@
  */
 package org.apache.hadoop.hbase.replication.regionserver;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyInt;
@@ -41,7 +41,6 @@ import org.apache.hadoop.hbase.Cell;
 import org.apache.hadoop.hbase.Cell.Type;
 import org.apache.hadoop.hbase.CellBuilderFactory;
 import org.apache.hadoop.hbase.CellBuilderType;
-import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.HRegionLocation;
@@ -73,13 +72,12 @@ import org.apache.hadoop.hbase.wal.WAL.Entry;
 import org.apache.hadoop.hbase.wal.WALEdit;
 import org.apache.hadoop.hbase.wal.WALKeyImpl;
 import org.apache.hadoop.hbase.zookeeper.ZKConfig;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.rules.TestName;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -90,12 +88,9 @@ import org.apache.hbase.thirdparty.com.google.common.util.concurrent.Uninterrupt
  * Tests RegionReplicaReplicationEndpoint class by setting up region replicas and verifying async
  * wal replication replays the edits to the secondary region in various scenarios.
  */
-@Category({ FlakeyTests.class, LargeTests.class })
+@Tag(FlakeyTests.TAG)
+@Tag(LargeTests.TAG)
 public class TestRegionReplicaReplicationEndpoint {
-
-  @ClassRule
-  public static final HBaseClassTestRule CLASS_RULE =
-    HBaseClassTestRule.forClass(TestRegionReplicaReplicationEndpoint.class);
 
   private static final Logger LOG =
     LoggerFactory.getLogger(TestRegionReplicaReplicationEndpoint.class);
@@ -104,10 +99,9 @@ public class TestRegionReplicaReplicationEndpoint {
 
   private static final HBaseTestingUtility HTU = new HBaseTestingUtility();
 
-  @Rule
-  public TestName name = new TestName();
+  private String methodName;
 
-  @BeforeClass
+  @BeforeAll
   public static void beforeClass() throws Exception {
     Configuration conf = HTU.getConfiguration();
     conf.setFloat("hbase.regionserver.logroll.multiplier", 0.0003f);
@@ -127,9 +121,14 @@ public class TestRegionReplicaReplicationEndpoint {
     HTU.startMiniCluster(NB_SERVERS);
   }
 
-  @AfterClass
+  @AfterAll
   public static void afterClass() throws Exception {
     HTU.shutdownMiniCluster();
+  }
+
+  @BeforeEach
+  public void before(TestInfo testInfo) {
+    methodName = testInfo.getTestMethod().get().getName();
   }
 
   @Test
@@ -325,7 +324,7 @@ public class TestRegionReplicaReplicationEndpoint {
   @Test
   public void testRegionReplicaWithoutMemstoreReplication() throws Exception {
     int regionReplication = 3;
-    final TableName tableName = TableName.valueOf(name.getMethodName());
+    final TableName tableName = TableName.valueOf(methodName);
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     htd.setRegionReplication(regionReplication);
     htd.setRegionMemstoreReplication(false);
@@ -361,7 +360,7 @@ public class TestRegionReplicaReplicationEndpoint {
     // does not test whether the replicas actually pick up flushed files and apply compaction
     // to their stores
     int regionReplication = 3;
-    final TableName tableName = TableName.valueOf(name.getMethodName());
+    final TableName tableName = TableName.valueOf(methodName);
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     htd.setRegionReplication(regionReplication);
     createOrEnableTableWithRetries(htd, true);
@@ -408,8 +407,8 @@ public class TestRegionReplicaReplicationEndpoint {
     // tests having edits from a disabled or dropped table is handled correctly by skipping those
     // entries and further edits after the edits from dropped/disabled table can be replicated
     // without problems.
-    final TableName tableName = TableName.valueOf(
-      name.getMethodName() + "_drop_" + dropTable + "_disabledReplication_" + disableReplication);
+    final TableName tableName = TableName
+      .valueOf(methodName + "_drop_" + dropTable + "_disabledReplication_" + disableReplication);
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     int regionReplication = 3;
     htd.setRegionReplication(regionReplication);
@@ -515,7 +514,7 @@ public class TestRegionReplicaReplicationEndpoint {
 
   @Test
   public void testMetaCacheMissTriggersRefresh() throws Exception {
-    TableName tableName = TableName.valueOf(name.getMethodName());
+    TableName tableName = TableName.valueOf(methodName);
     int regionReplication = 3;
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     htd.setRegionReplication(regionReplication);
@@ -555,8 +554,7 @@ public class TestRegionReplicaReplicationEndpoint {
       sinkWriter.append(tableName, encodedRegionName, Bytes.toBytes("testRow"),
         Lists.newArrayList(entry));
 
-      assertEquals("No entries should be skipped for valid table", 0, skippedEdits.get());
-
+      assertEquals(0, skippedEdits.get(), "No entries should be skipped for valid table");
     } finally {
       table.close();
       connection.close();
@@ -565,7 +563,7 @@ public class TestRegionReplicaReplicationEndpoint {
 
   @Test
   public void testNullTableDescriptorDoesNotCauseNPE() throws Exception {
-    TableName tableName = TableName.valueOf(name.getMethodName());
+    TableName tableName = TableName.valueOf(methodName);
     int regionReplication = 2;
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     htd.setRegionReplication(regionReplication);
@@ -618,7 +616,7 @@ public class TestRegionReplicaReplicationEndpoint {
 
   @Test
   public void testMetaCacheSkippedForSingleReplicaTable() throws Exception {
-    TableName tableName = TableName.valueOf(name.getMethodName());
+    TableName tableName = TableName.valueOf(methodName);
     int regionReplication = 1;
     HTableDescriptor htd = HTU.createTableDescriptor(tableName);
     htd.setRegionReplication(regionReplication);
@@ -658,8 +656,7 @@ public class TestRegionReplicaReplicationEndpoint {
       sinkWriter.append(tableName, encodedRegionName, Bytes.toBytes("testRow"),
         Lists.newArrayList(entry));
 
-      assertEquals("No entries should be skipped for single replica table", 0, skippedEdits.get());
-
+      assertEquals(0, skippedEdits.get(), "No entries should be skipped for single replica table");
     } finally {
       table.close();
       connection.close();
