+++ hbase-server/src/test/java/org/apache/hadoop/hbase/replication/regionserver/TestReplicationMarker.java
@@ -25,9 +25,9 @@ import static org.apache.hadoop.hbase.replication.master.ReplicationSinkTrackerT
 import static org.apache.hadoop.hbase.replication.master.ReplicationSinkTrackerTableCreator.WAL_NAME_COLUMN;
 import static org.apache.hadoop.hbase.replication.regionserver.ReplicationMarkerChore.REPLICATION_MARKER_CHORE_DURATION_KEY;
 import static org.apache.hadoop.hbase.replication.regionserver.ReplicationMarkerChore.REPLICATION_MARKER_ENABLED_KEY;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -35,7 +35,6 @@ import java.util.List;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.Cell;
 import org.apache.hadoop.hbase.CellUtil;
-import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
 import org.apache.hadoop.hbase.HConstants;
@@ -53,11 +52,10 @@ import org.apache.hadoop.hbase.testclassification.MediumTests;
 import org.apache.hadoop.hbase.testclassification.ReplicationTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.wal.WAL;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -69,11 +67,9 @@ import org.slf4j.LoggerFactory;
  * "hbase.regionserver.replication.sink.tracker.enabled" conf key enabled. This will persist the
  * marker rows coming from peer cluster to persist to REPLICATION.SINK_TRACKER table.
  **/
-@Category({ ReplicationTests.class, MediumTests.class })
+@Tag(ReplicationTests.TAG)
+@Tag(MediumTests.TAG)
 public class TestReplicationMarker {
-  @ClassRule
-  public static final HBaseClassTestRule CLASS_RULE =
-    HBaseClassTestRule.forClass(TestReplicationMarker.class);
 
   private static final Logger LOG = LoggerFactory.getLogger(TestReplicationMarker.class);
 
@@ -82,7 +78,7 @@ public class TestReplicationMarker {
   private static HBaseTestingUtility utility1;
   private static HBaseTestingUtility utility2;
 
-  @BeforeClass
+  @BeforeAll
   public static void setUpBeforeClass() throws Exception {
     conf1 = HBaseConfiguration.create();
     conf1.set(HConstants.ZOOKEEPER_ZNODE_PARENT, "/1");
@@ -111,15 +107,15 @@ public class TestReplicationMarker {
     ReplicationSourceManager manager = utility1.getHBaseCluster().getRegionServer(0)
       .getReplicationSourceService().getReplicationManager();
     // Wait until the peer gets established.
-    Waiter.waitFor(conf1, 10000, (Waiter.Predicate) () -> manager.getSources().size() == 1);
+    Waiter.waitFor(conf1, 10000, () -> manager.getSources().size() == 1);
   }
 
   private static void waitForReplicationTrackerTableCreation() {
-    Waiter.waitFor(conf2, 10000, (Waiter.Predicate) () -> utility2.getAdmin()
-      .tableExists(REPLICATION_SINK_TRACKER_TABLE_NAME));
+    Waiter.waitFor(conf2, 10000,
+      () -> utility2.getAdmin().tableExists(REPLICATION_SINK_TRACKER_TABLE_NAME));
   }
 
-  @AfterClass
+  @AfterAll
   public static void tearDown() throws Exception {
     utility1.shutdownMiniCluster();
     utility2.shutdownMiniCluster();
@@ -131,14 +127,14 @@ public class TestReplicationMarker {
     // create enough sentinel rows.
     Thread.sleep(5000);
     WAL wal1 = utility1.getHBaseCluster().getRegionServer(0).getWAL(null);
-    String walName1ForCluster1 = ((AbstractFSWAL) wal1).getCurrentFileName().getName();
+    String walName1ForCluster1 = ((AbstractFSWAL<?>) wal1).getCurrentFileName().getName();
     String rs1Name = utility1.getHBaseCluster().getRegionServer(0).getServerName().getHostname();
     // Since we sync the marker edits while appending to wal, all the edits should be visible
     // to Replication threads immediately.
     assertTrue(getReplicatedEntries() >= 5);
     // Force log roll.
     wal1.rollWriter(true);
-    String walName2ForCluster1 = ((AbstractFSWAL) wal1).getCurrentFileName().getName();
+    String walName2ForCluster1 = ((AbstractFSWAL<?>) wal1).getCurrentFileName().getName();
     Connection connection2 = utility2.getMiniHBaseCluster().getRegionServer(0).getConnection();
     // Sleep for 5 more seconds to get marker rows with new wal name.
     Thread.sleep(5000);
