+++ hbase-server/src/test/java/org/apache/hadoop/hbase/replication/regionserver/TestSerialReplicationEndpoint.java
@@ -17,6 +17,9 @@
  */
 package org.apache.hadoop.hbase.replication.regionserver;
 
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -26,7 +29,6 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.LinkedBlockingQueue;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.Cell;
-import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseTestingUtility;
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.ServerName;
@@ -46,29 +48,24 @@ import org.apache.hadoop.hbase.testclassification.ReplicationTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.EnvironmentEdgeManager;
 import org.apache.hadoop.hbase.wal.WAL.Entry;
-import org.junit.AfterClass;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
 
 import org.apache.hbase.thirdparty.com.google.common.collect.ImmutableList;
 import org.apache.hbase.thirdparty.com.google.common.collect.ImmutableMap;
 import org.apache.hbase.thirdparty.com.google.common.io.Closeables;
 
-@Category({ ReplicationTests.class, MediumTests.class })
+@Tag(ReplicationTests.TAG)
+@Tag(MediumTests.TAG)
 public class TestSerialReplicationEndpoint {
-
-  @ClassRule
-  public static final HBaseClassTestRule CLASS_RULE =
-    HBaseClassTestRule.forClass(TestSerialReplicationEndpoint.class);
 
   private static HBaseTestingUtility UTIL = new HBaseTestingUtility();
   private static Configuration CONF;
   private static Connection CONN;
 
-  @BeforeClass
+  @BeforeAll
   public static void setUp() throws Exception {
     UTIL.startMiniCluster();
     CONF = UTIL.getConfiguration();
@@ -76,7 +73,7 @@ public class TestSerialReplicationEndpoint {
     CONN = UTIL.getConnection();
   }
 
-  @AfterClass
+  @AfterAll
   public static void tearDown() throws Exception {
     Closeables.close(CONN, true);
     UTIL.shutdownMiniCluster();
@@ -118,7 +115,7 @@ public class TestSerialReplicationEndpoint {
     Waiter.waitFor(CONF, 60000, () -> TestEndpoint.getEntries().size() >= cellNum);
 
     int index = 0;
-    Assert.assertEquals(TestEndpoint.getEntries().size(), cellNum);
+    assertEquals(cellNum, TestEndpoint.getEntries().size());
     if (!isSerial) {
       Collections.sort(TestEndpoint.getEntries(), (a, b) -> {
         long seqA = a.getKey().getSequenceId();
@@ -127,15 +124,14 @@ public class TestSerialReplicationEndpoint {
       });
     }
     for (Entry entry : TestEndpoint.getEntries()) {
-      Assert.assertEquals(entry.getKey().getTableName(), tableName);
-      Assert.assertEquals(entry.getEdit().getCells().size(), 1);
+      assertEquals(tableName, entry.getKey().getTableName());
+      assertEquals(1, entry.getEdit().getCells().size());
       Cell cell = entry.getEdit().getCells().get(0);
-      Assert.assertArrayEquals(
-        Bytes.copy(cell.getRowArray(), cell.getRowOffset(), cell.getRowLength()),
-        Bytes.toBytes(index));
+      assertArrayEquals(Bytes.toBytes(index),
+        Bytes.copy(cell.getRowArray(), cell.getRowOffset(), cell.getRowLength()));
       index++;
     }
-    Assert.assertEquals(index, cellNum);
+    assertEquals(cellNum, index);
   }
 
   @Test
