+++ hbase-server/src/test/java/org/apache/hadoop/hbase/replication/regionserver/TestWALEntrySinkFilter.java
@@ -17,8 +17,8 @@
  */
 package org.apache.hadoop.hbase.replication.regionserver;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -26,7 +26,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hbase.Cell;
@@ -34,11 +33,9 @@ import org.apache.hadoop.hbase.CellBuilder;
 import org.apache.hadoop.hbase.CellBuilderFactory;
 import org.apache.hadoop.hbase.CellBuilderType;
 import org.apache.hadoop.hbase.CellScanner;
-import org.apache.hadoop.hbase.HBaseClassTestRule;
 import org.apache.hadoop.hbase.HBaseConfiguration;
 import org.apache.hadoop.hbase.HConstants;
 import org.apache.hadoop.hbase.ServerName;
-import org.apache.hadoop.hbase.Stoppable;
 import org.apache.hadoop.hbase.TableName;
 import org.apache.hadoop.hbase.client.AdvancedScanResultConsumer;
 import org.apache.hadoop.hbase.client.AsyncAdminBuilder;
@@ -59,13 +56,9 @@ import org.apache.hadoop.hbase.security.User;
 import org.apache.hadoop.hbase.testclassification.ReplicationTests;
 import org.apache.hadoop.hbase.testclassification.SmallTests;
 import org.apache.hadoop.hbase.util.Bytes;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.rules.TestName;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 
 import org.apache.hbase.thirdparty.com.google.protobuf.ByteString;
 
@@ -74,37 +67,13 @@ import org.apache.hadoop.hbase.shaded.protobuf.generated.AdminProtos;
 /**
  * Simple test of sink-side wal entry filter facility.
  */
-@Category({ ReplicationTests.class, SmallTests.class })
+@Tag(ReplicationTests.TAG)
+@Tag(SmallTests.TAG)
 public class TestWALEntrySinkFilter {
 
-  @ClassRule
-  public static final HBaseClassTestRule CLASS_RULE =
-    HBaseClassTestRule.forClass(TestWALEntrySinkFilter.class);
-
-  private static final Logger LOG = LoggerFactory.getLogger(TestWALEntrySinkFilter.class);
-  @Rule
-  public TestName name = new TestName();
   static final int BOUNDARY = 5;
   static final AtomicInteger UNFILTERED = new AtomicInteger();
   static final AtomicInteger FILTERED = new AtomicInteger();
-
-  /**
-   * Implemetentation of Stoppable to pass into ReplicationSink.
-   */
-  private static Stoppable STOPPABLE = new Stoppable() {
-    private final AtomicBoolean stop = new AtomicBoolean(false);
-
-    @Override
-    public boolean isStopped() {
-      return this.stop.get();
-    }
-
-    @Override
-    public void stop(String why) {
-      LOG.info("STOPPING BECAUSE: " + why);
-      this.stop.set(true);
-    }
-  };
 
   /**
    * Test filter. Filter will filter out any write time that is <= 5 (BOUNDARY). We count how many
@@ -113,7 +82,8 @@ public class TestWALEntrySinkFilter {
    * our counting Table.
    */
   @Test
-  public void testWALEntryFilter() throws IOException {
+  public void testWALEntryFilter(TestInfo testInfo) throws IOException {
+    String testName = testInfo.getTestMethod().get().getName();
     Configuration conf = HBaseConfiguration.create();
     // Make it so our filter is instantiated on construction of ReplicationSink.
     conf.setClass(HConstants.CLIENT_CONNECTION_REGISTRY_IMPL_CONF_KEY,
@@ -128,8 +98,7 @@ public class TestWALEntrySinkFilter {
       new ArrayList<>();
     AdminProtos.WALEntry.Builder entryBuilder = AdminProtos.WALEntry.newBuilder();
     // Need a tablename.
-    ByteString tableName =
-      ByteString.copyFromUtf8(TableName.valueOf(this.name.getMethodName()).toString());
+    ByteString tableName = ByteString.copyFromUtf8(TableName.valueOf(testName).toString());
     // Add WALEdit Cells to Cells List. The way edits arrive at the sink is with protos
     // describing the edit with all Cells from all edits aggregated in a single CellScanner.
     final List<Cell> cells = new ArrayList<>();
@@ -294,7 +263,7 @@ public class TestWALEntrySinkFilter {
           for (Row action : actions) {
             // Row is the index of the loop above where we make WALEntry and Cells.
             int row = Bytes.toInt(action.getRow());
-            assertTrue("" + row, row > BOUNDARY);
+            assertTrue(row > BOUNDARY, "" + row);
             UNFILTERED.incrementAndGet();
             list.add(null);
           }
